--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -8,9 +8,7 @@ require (
 )
 
 require (
-	github.com/aclements/go-moremath v0.0.0-20210112150236-f10218a38794 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	golang.org/x/perf v0.0.0-20231006134539-cd219cffda85 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )