--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"fmt"
 	"io"
-	"regexp"
 	"strings"
 
 	"github.com/goccy/go-json"
@@ -23,42 +22,25 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
 	var user User
-	var i int
+	countResults := make(DomainStat, bufio.MaxScanTokenSize/2)
+	suffix := "." + domain
+
 	scanner := bufio.NewScanner(r)
 	scanner.Split(bufio.ScanLines)
-	for scanner.Scan() {
-		if err = json.Unmarshal(scanner.Bytes(), &user); err != nil {
-			return
-		}
-		result[i] = user
-		i++
-	}
-	return
-}
 
-func countDomains(u users, domain string) (DomainStat, error) {
-	result := make(DomainStat, len(u)/6)
-	rc, err := regexp.Compile("\\." + domain)
-	if err != nil {
-		return nil, err
-	}
-	for _, user := range u {
-		if matched := rc.Match([]byte(user.Email)); matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+	for scanner.Scan() {
+		if err := json.Unmarshal(scanner.Bytes(), &user); err != nil {
+			return countResults, fmt.Errorf("get users error: %w", err)
+		}
+		if user.Email != "" {
+			if matched := strings.HasSuffix(user.Email, suffix); matched {
+				resultString := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+				num := countResults[resultString]
+				num++
+				countResults[resultString] = num
+			}
 		}
 	}
-	return result, nil
+	return countResults, nil
 }