--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -10,13 +10,7 @@ import (
 )
 
 type User struct {
-	ID       int
-	Name     string
-	Username string
-	Email    string
-	Phone    string
-	Password string
-	Address  string
+	Email string
 }
 
 type DomainStat map[string]int
@@ -33,12 +27,14 @@ func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 		if err := json.Unmarshal(scanner.Bytes(), &user); err != nil {
 			return countResults, fmt.Errorf("get users error: %w", err)
 		}
+
 		if user.Email != "" {
-			if matched := strings.HasSuffix(user.Email, suffix); matched {
-				resultString := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
-				num := countResults[resultString]
-				num++
-				countResults[resultString] = num
+			if strings.HasSuffix(user.Email, suffix) {
+				email := strings.SplitN(user.Email, "@", 2)
+				if len(email) == 2 {
+					resultString := strings.ToLower(email[1])
+					countResults[resultString]++
+				}
 			}
 		}
 	}