## Домашнее задание №10 «Оптимизация программы»
Benchmark command:
``` go test -bench=BenchmarkGetDomainStat -benchmem -benchtime 10s -count 10 ```

Benchstat comparison results:
```
goos: linux
goarch: amd64
pkg: github.com/truep/golan-otus-hw/hw10_program_optimization
cpu: AMD Ryzen 7 5800H with Radeon Graphics         
                 │   old.out    │                new.out                │              new10.out              │
                 │    sec/op    │    sec/op     vs base                 │   sec/op     vs base                │
GetDomainStat-16   1281.6µ ± 6%   3882.1µ ± 2%  +202.91% (p=0.000 n=10)   213.8µ ± 5%  -83.32% (p=0.000 n=10)

                 │   old.out    │                 new.out                 │                new10.out                 │
                 │     B/op     │     B/op       vs base                  │      B/op       vs base                  │
GetDomainStat-16   35.17Ki ± 5%   900.42Ki ± 0%  +2460.09% (p=0.000 n=10)   1772.13Ki ± 0%  +4938.55% (p=0.000 n=10)

                 │   old.out    │               new.out                │             new10.out              │
                 │  allocs/op   │  allocs/op    vs base                │ allocs/op   vs base                │
GetDomainStat-16   348.000 ± 5%   101.000 ± 5%  -70.98% (p=0.000 n=10)   6.000 ± 0%  -98.28% (p=0.000 n=10)


```

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов